* Massive changes to the way the project explorer deals with its nodes, it now works based on a tree structure. Child nodes can build their file paths dynamically based on their parent link.
The absolute path is determined using recursion. Each parent calls the
next parent up in the chain. When the the call has reached the root of
the tree the path is complete.

When a node is moved all it needs to do is update its parent link, this
"linked tree" structure that I designed is extremely flexible and will be
ported over to other editors.

* All of the known bugs in the ProjectExplorer with regard to renaming/moving
folders when child documents are opened in the code editor have been
fixed/hacked.

TBD:
The code in the ProjectExplorer.cs file is horrible, the Model-View-Controller
are completely blurred, coupling is high making it hard to hunt down bugs.
Will be refactored before release of version 1.0.

Signed-off-by: swordmaster2k <b00056835@student.itb.ie>